requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

cryptography==43.0.3
azure-functions
azure-storage-blob
azure-storage-queue
pyodbc
aiohttp
aiodns
typing-extensions
pymssql